requirements.txt
streamlit>=1.28.0
google-genai==1.3.0
pillow>=9.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0 
